Fix issue #214, don't ask for backup if restoring from backup
This was working correctly for file restores, but not ipfs.

Changed importipfs function to handle the vault string the same as
reasSingleFile. That is, pass the string into _parseVault, not the
parsed object.

This will allow the _parseVault function to parse the serialized vault
object and calling "setvaultDownloaded" if this is a V1Backup.